+++ lib/methods_functions/general_methods.dart
@@ -12,8 +12,8 @@ import 'package:archive/archive_io.dart';
 
 //Import alert dialogs and others
 import "package:simple_chat/utils/alert_dialog_list.dart";
-import 'package:simple_chat/configurations/config_invoke.dart';
-import "package:simple_chat/classes/classes.dart";
+import 'package:simple_chat/configuration/config_invoke.dart';
+import "package:simple_chat/message_related/message_class.dart";
 import 'package:simple_chat/utils/colorimetry_blueprints.dart';
 import 'package:simple_chat/utils/logger_config.dart';
