+++ lib/session_related/refresh_tk_watch_dog.dart
@@ -1,7 +1,7 @@
 import 'dart:async';
 import 'package:flutter/cupertino.dart';
-import 'package:simple_chat/classes/app_storage.dart';
-import 'package:simple_chat/configurations/config_invoke.dart';
+import 'package:simple_chat/session_related/app_storage_class.dart';
+import 'package:simple_chat/configuration/config_invoke.dart';
 import 'package:simple_chat/utils/logger_config.dart';
 
 import 'package:simple_chat/methods_functions/user_entrypoint_methods.dart';
